kv: fix the rope tail read in attn_comp_pack_ld (the canonical loader)
e09563e narrowed the ATTN_PACK rope tail to bf16 and updated every reader I
could find -- except the one that matters most. attn_comp_pack_ld in
pulsar_cuda_internal.h:139 still read the tail through `const float *`, so it
consumed 4 bytes where the row now holds 2: it fused two bf16 lanes into one
bogus f32, then walked off the end of the row into the next row's nope bytes.
Generation collapsed to repeated <BOS> at any context depth.

The miss is instructive. pulsar_cuda_attention.cu:48 claims attn_pack_e4m3 and
attn_comp_pack_ld "now live in pulsar_cuda_internal.h so the fp16 tensor-core
kernel decodes ATTN_PACK rows with the SAME code rather than a transcription of
it -- there is then no second copy of the contract to drift or to get wrong."
That claim is false: pulsar_cuda_attn_f16.cu:325-341 IS a second copy, an inline
float4 path that never calls the shared loader. I found and fixed the copy,
which read as if it were the original, and left the original untouched.

WHY EVERY GATE STAYED GREEN. attn_decode_split compares a golden arm against a
split arm and attn_f16_banked compares packed addressing against a constant --
but both arms decode through attn_comp_pack_ld, so a defect INSIDE the shared
loader cancels on both sides and the differential reads clean. The packed cases
even scored 1.6e-10, tighter than the f32 path, which is exactly what a
consistently-wrong-on-both-sides comparison looks like. Same trap as the SoA
misattribution: a differential harness cannot see a bug in code both arms share.
Only end-to-end generation caught it, and only because the output was obviously
garbage rather than subtly worse.

Every rope accessor is now bf16, verified exhaustively rather than by memory:
the 3 nblk_pad sites in pulsar_cuda_norm_kv.cu (store/dequant/repack), the 4
SCALES_PAD sites in pulsar_cuda_attention.cu (scalar dot, lane8 dot, both float4
smem staging paths), attn_f16.cu's transcription, the shared loader itself, and
the 2 gates that synthesise rows by hand.

+++ src/cuda/pulsar_cuda_internal.h
@@ -136,7 +136,8 @@ __device__ static inline float attn_comp_pack_ld(const float *comp_kv, uint64_t 
         const float scale = __uint_as_float((uint32_t)r[n_nope + (d / PULSAR_FP8_KV_BLOCK)] << 23);
         return attn_pack_e4m3(r[d], scale);
     }
-    return ((const float *)(r + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim)))[d - n_nope];
+    return __bfloat162float(((const __nv_bfloat16 *)(r + n_nope +
+                             PULSAR_ATTN_PACK_SCALES_PAD(head_dim)))[d - n_nope]);
 }
 
 struct pulsar_gpu_tensor {
